V4L2: OMAP34XX: Set ISPCTRL_SYNC_DETECT correctly
SYNC_DETECT bits in ISPCTRL register were defined incorrectly.
This leads to resetting of two lowest bits in the register,
disabling the reception of image data by the ISP. This patch
fixes the problem.

The bug was found by Tuukka Toivonen <tuukka.o.toivonen@nokia.com>.

+++ drivers/media/video/isp/isp.h
@@ -175,7 +175,7 @@ struct isp_interface_config {
 	u8 par_bridge;
 	u8 para_clk_pol;
 	u8 dataline_shift;
-	u8 hsvs_syncdetect;
+	u32 hsvs_syncdetect;
 	u16 vdint0_timing;
 	u16 vdint1_timing;
 	int strobe;
